Enable items to have 0 options if the item is hidden (for admin drafting purposes)

+++ services/MerchStoreService.ts
@@ -103,12 +103,12 @@ export default class MerchStoreService {
     });
   }
 
-  public async findItemByUuid(uuid: Uuid): Promise<PublicMerchItem> {
+  public async findItemByUuid(uuid: Uuid, canSeeFullItem = false): Promise<PublicMerchItem> {
     const item = await this.transactions.readOnly(async (txn) => Repositories
       .merchStoreItem(txn)
       .findByUuid(uuid));
     if (!item) throw new NotFoundError('Merch item not found');
-    return item.getPublicMerchItem();
+    return item.getPublicMerchItem(canSeeFullItem);
   }
 
   public async createItem(item: MerchItem): Promise<MerchandiseItemModel> {
@@ -128,6 +128,7 @@ export default class MerchStoreService {
   /**
    * Verify that items have valid options. An item with variants disabled cannot have multiple
    * options, and an item with variants enabled cannot have multiple option types.
+   * @param item merch item
    */
   private static verifyItemHasValidOptions(item: MerchItem | MerchandiseItemModel) {
     if (!item.hasVariantsEnabled && item.options.length > 1) {
@@ -143,23 +144,36 @@ export default class MerchStoreService {
     return optionTypes.size > 1;
   }
 
+  /**
+   * Edits a merch item and its options, given the item edit.
+   * Item edits cannot add or remove item options - they can only edit existing options.
+   * If the visibility of the item is set to visible, then the item cannot have 0 options.
+   * @param uuid item uuid
+   * @param itemEdit item edits
+   * @returns edited item
+   */
   public async editItem(uuid: Uuid, itemEdit: MerchItemEdit): Promise<MerchandiseItemModel> {
     return this.transactions.readWrite(async (txn) => {
       const merchItemRepository = Repositories.merchStoreItem(txn);
       const item = await merchItemRepository.findByUuid(uuid);
       if (!item) throw new NotFoundError();
+      if (itemEdit.hidden === false && item.options.length === 0) {
+        throw new UserError('Item cannot be set to visible if it has 0 options.');
+      }
+
       const { options, collection: updatedCollection, ...changes } = itemEdit;
 
       if (options) {
-        const updatedOptionsByUuid = new Map(options.map((option) => [option.uuid, option]));
-        item.options.map((option) => {
-          if (!updatedOptionsByUuid.has(option.uuid)) return;
-          const updatedOption = updatedOptionsByUuid.get(option.uuid);
+        const optionUpdatesByUuid = new Map(options.map((option) => [option.uuid, option]));
+
+        item.options.map((currentOption) => {
+          if (!optionUpdatesByUuid.has(currentOption.uuid)) return;
+          const optionUpdate = optionUpdatesByUuid.get(currentOption.uuid);
           // 'quantity' is incremented instead of directly set to avoid concurrency issues with orders
           // e.g. there's 10 of an item and someone adds 5 to stock while someone else orders 1
           // so the merch store admin sets quantity to 15 but the true quantity is 14
-          if (updatedOption.addQuantity) updatedOption.addQuantity += option.quantity;
-          return MerchandiseItemOptionModel.merge(option, updatedOption);
+          if (optionUpdate.quantityToAdd) currentOption.quantity += optionUpdate.quantityToAdd;
+          return MerchandiseItemOptionModel.merge(currentOption, optionUpdate);
         });
       }
 
@@ -188,14 +202,29 @@ export default class MerchStoreService {
     });
   }
 
+  /**
+   * Creates an item option. An item option can be added to an item if:
+   *    - the item has variants enabled and the option has the same type as the existing item options
+   *    - the item has variants disabled and has exactly 0 options (only the case if the item is hidden)
+   * @param item merch item uuid
+   * @param option merch item option
+   * @returns created item option
+   */
   public async createItemOption(item: Uuid, option: MerchItemOption): Promise<PublicMerchItemOption> {
     return this.transactions.readWrite(async (txn) => {
       const merchItem = await Repositories.merchStoreItem(txn).findByUuid(item);
       if (!merchItem) throw new NotFoundError('Merch item not found');
-      if (!merchItem.hasVariantsEnabled) throw new UserError('Cannot add option to items with variants disabled');
-      // check only the first option since all other options must have same type as first based on prior validation
-      const hasDifferentOptionType = merchItem.options[0].metadata?.type !== option.metadata?.type;
-      if (hasDifferentOptionType) throw new UserError('Merch item cannot have multiple option types');
+      if (!merchItem.hasVariantsEnabled && merchItem.options.length > 0) {
+        throw new UserError('Cannot add more than 1 option to items with variants disabled');
+      }
+
+      // Check that only the first option's type matches the option type to be added,
+      // since all other options must have same type as the first, based on prior validation.
+      //
+      // Note: options[0] could be undefined here since an item is allowed to have 0 options if it is hidden.
+      const hasDifferentOptionTypeThanOtherOptions = merchItem.hasVariantsEnabled
+          && merchItem.options[0]?.metadata.type !== option.metadata.type;
+      if (hasDifferentOptionTypeThanOtherOptions) throw new UserError('Merch item cannot have multiple option types');
 
       const merchItemOptionRepository = Repositories.merchStoreItemOption(txn);
       const createdOption = MerchandiseItemOptionModel.create({ ...option, item: merchItem });
@@ -204,6 +233,13 @@ export default class MerchStoreService {
     });
   }
 
+  /**
+   * Deletetes the given item option. Deletion will fail if the item option has already been ordered,
+   * or if the deletion will result in the item having 0 options while being visible to the public.
+   *
+   * Note that the item is allowed to have 0 options, but only if the item is hidden.
+   * @param uuid option uuid
+   */
   public async deleteItemOption(uuid: Uuid): Promise<void> {
     await this.transactions.readWrite(async (txn) => {
       const merchItemOptionRepository = Repositories.merchStoreItemOption(txn);
@@ -211,6 +247,10 @@ export default class MerchStoreService {
       if (!option) throw new NotFoundError();
       const hasBeenOrdered = await Repositories.merchOrderItem(txn).hasOptionBeenOrdered(uuid);
       if (hasBeenOrdered) throw new UserError('This item option has been ordered and cannot be deleted');
+      if (option.item.options.length === 1 && !option.item.hidden) {
+        throw new UserError('Cannot delete the only option for an item when it is visible');
+      }
+
       return merchItemOptionRepository.deleteMerchItemOption(option);
     });
   }
@@ -234,6 +274,20 @@ export default class MerchStoreService {
     return orders.map((o) => o.getPublicOrder());
   }
 
+  /**
+   * Places an order with the list of options and their quantities for the given user.
+   *
+   * The order is placed if the following conditions are met:
+   *    - all the ordered item options exist within the database
+   *    - the ordered item options were placed for non-hidden items
+   *    - the user wouldn't reach monthly and lifetime limits for the item if this order is placed
+   *    - the requested item options are in stock
+   *    - the user has enough credits to place the order
+   *
+   * @param originalOrder the order containing item options and their quantities
+   * @param user user placing the order
+   * @returns PublicOrder object placed
+   */
   public async placeOrder(originalOrder: MerchItemOptionAndQuantity[], user: UserModel): Promise<PublicOrder> {
     const [order, merchItemOptions] = await this.transactions.readWrite(async (txn) => {
       await user.reload();
@@ -245,7 +299,7 @@ export default class MerchStoreService {
           .filter((o) => !o.item.hidden)
           .map((o) => o.uuid);
         const missingItems = difference(requestedItems, foundItems);
-        throw new NotFoundError(`Missing: ${missingItems}`);
+        throw new NotFoundError(`The following items were not found: ${missingItems}`);
       }
 
       // Checks that hidden items were not ordered
